+++ Swich/ItemDAO.h
@@ -52,7 +52,7 @@ public:
   void setNewTable(QString name, QString password);
   void deleteTable(QString tableName);
 
-  void addItemToTable(QString table, QString name, int buyP, int sellP);
+  void addItemToTable(QString table, QString name, double buyP, double sellP);
   void deleteItemToTable(QString tableName, QString itemName);
 
 private:
